✨feat: 결제 영수증 모델 및 CoreData 생성 ♻️refactor: 기존 CoreData 코드 수정 및 연계 코드 수정
feat:
- 결졔 영수증 모델 제작
- 결제 영수증 CoreData 생성
- CoreDataManager에 결제 영수증 CRUD 구현

refactor:
- animal CoreData CRUD 수정
- LocationViewModel 에서 animalEntity Read 연동 메소드 수정

+++ Animality/Animality/Model/RentReceiptModel.swift
@@ -18,7 +18,7 @@ struct RentReceipt {
     /// 대여한 동물
     let animalId: UUID
     /// 결제 금액
-    let amount: Int
+    let amount: Int64
     /// 대여 장소
     let location: String?
     /// 결제 시간
